removed unnecessary reducer
